requirements.txt

streamlit>=1.36
pydantic>=2.7
python-dateutil>=2.9
pyyaml>=6.0.1
pandas>=2.2
